Add own recipe to DB

+++ code/Reciply/Reciply/Models/Recipe.cs
@@ -20,7 +20,7 @@ namespace Reciply.Models
             }
         } //The Ingredients 1/n
         public int Portion { get; set; } //One Portion = food for one people
-        public int Rating { get; set; } //How many stars the food is rated 
+        public double Rating { get; set; } //How many stars the food is rated 
         public double Duration { get; set; } //How long does the recipe need to be cooked
         public string Tags { get; set; } //Which tags the food has for example
         public string Description { get; set; } //The description of the food for example.: Carbonara is food with bacon and pasta
